chore: Cleanup everything and add inline comments (#34)
* chore: Cleanup everything and add inline comments

- int256 -> uint256 contract size in Proxy
- removed deprecated `PROXY_CODE_HASH` from Factory
- made `_initializeInstance` private and only called internally, so no need to check proxy contract size
- moved reused code into a private `_getImplementationOfProxy` function that also does not need to check proxy contract size
- `_initializeInstance` and `_upgradeInstance` now return false if no initializer/migrator but has arguments
- all deploys and calls low-level in order to not have reverts and bubble up all successes to inheritor
- cleaned up logic and optimized
- version 0 now explicitly banned
- added `_setFactory` to `Proxied`

* chore: natspec and comments

+++ contracts/Proxy.sol
@@ -28,13 +28,13 @@ contract Proxy is SlotManipulatable {
     fallback() payable external virtual {
         bytes32 implementation = _getSlotValue(IMPLEMENTATION_SLOT);
 
-        int256 size;
+        uint256 size;
 
         assembly {
             size := extcodesize(implementation)
         }
 
-        require(size != 0, "P:F:INVALID_IMPLEMENTATION");
+        require(size != uint256(0), "P:F:INVALID_IMPLEMENTATION");
 
         assembly {
             calldatacopy(0, 0, calldatasize())
